webkit: Refactor timezone handling to be a context feature
Previously the timezone override handling was exposed through the Inspector Page
agent, but now it is moving to the Playwright agent which will hopefully rely on
the new WebKit upstream PageConfiguration API for this.

This patch doesn't change the Playwright top-level API. It impacts only the
underlying WebKit browser support.